+++ src/main/java/com/robertx22/age_of_exile/aoe_data/database/perks/GameChangerPerks.java
@@ -5,6 +5,7 @@ import com.robertx22.age_of_exile.database.OptScaleExactStat;
 import com.robertx22.age_of_exile.database.data.stats.types.defense.Armor;
 import com.robertx22.age_of_exile.database.data.stats.types.defense.DodgeRating;
 import com.robertx22.age_of_exile.database.data.stats.types.generated.ElementalDamageBonus;
+import com.robertx22.age_of_exile.database.data.stats.types.offense.DotDamage;
 import com.robertx22.age_of_exile.database.data.stats.types.offense.NonCritDamage;
 import com.robertx22.age_of_exile.database.data.stats.types.offense.TotalDamage;
 import com.robertx22.age_of_exile.database.data.stats.types.offense.crit.CriticalDamage;
@@ -60,7 +61,7 @@ public class GameChangerPerks implements ISlashRegistryInit {
         );
 
         PerkBuilder.gameChanger("steady_hand", "Steady Hand",
-            new OptScaleExactStat(-500, CriticalDamage.getInstance(), ModType.GLOBAL_INCREASE),
+            new OptScaleExactStat(-100, CriticalDamage.getInstance(), ModType.GLOBAL_INCREASE),
             new OptScaleExactStat(20, TotalDamage.getInstance(), ModType.FLAT)
         );
 
@@ -93,6 +94,11 @@ public class GameChangerPerks implements ISlashRegistryInit {
         PerkBuilder.gameChanger("divinity", "Divinity",
             new OptScaleExactStat(25, DatapackStatAdder.HEAL_TO_SPELL_DMG, ModType.FLAT),
             new OptScaleExactStat(-50, CriticalDamage.getInstance(), ModType.GLOBAL_INCREASE)
+        );
+
+        PerkBuilder.gameChanger("tormentor", "Tormentor",
+            new OptScaleExactStat(35, DotDamage.getInstance(), ModType.FLAT),
+            new OptScaleExactStat(-10, TotalDamage.getInstance(), ModType.FLAT)
         );
 
     }