+++ src/main/java/com/robertx22/age_of_exile/database/data/spells/components/selectors/AoeSelector.java
@@ -2,6 +2,8 @@ package com.robertx22.age_of_exile.database.data.spells.components.selectors;
 
 import com.robertx22.age_of_exile.database.data.spells.components.MapHolder;
 import com.robertx22.age_of_exile.database.data.spells.components.tooltips.ICTextTooltip;
+import com.robertx22.age_of_exile.database.data.spells.spell_classes.SpellCtx;
+import com.robertx22.age_of_exile.database.data.spells.spell_classes.SpellModEnum;
 import com.robertx22.age_of_exile.saveclasses.gearitem.gear_bases.TooltipInfo;
 import com.robertx22.age_of_exile.uncommon.utilityclasses.EntityFinder;
 import com.robertx22.age_of_exile.uncommon.utilityclasses.RandomUtils;
@@ -57,9 +59,11 @@ public class AoeSelector extends BaseTargetSelector implements ICTextTooltip {
     }
 
     @Override
-    public List<LivingEntity> get(LivingEntity caster, LivingEntity target, BlockPos pos, MapHolder data) {
+    public List<LivingEntity> get(SpellCtx ctx, LivingEntity caster, LivingEntity target, BlockPos pos, MapHolder data) {
         EntityFinder.EntityPredicate predicate = data.getEntityPredicate();
         Double radius = data.get(RADIUS);
+
+        radius *= ctx.calculatedSpellData.config.getMulti(SpellModEnum.AREA);
 
         Double chance = data.getOrDefault(SELECTION_CHANCE, 100D);