+++ src/main/java/com/robertx22/age_of_exile/database/data/spells/components/selectors/BaseTargetSelector.java
@@ -4,6 +4,7 @@ import com.robertx22.age_of_exile.database.data.IGUID;
 import com.robertx22.age_of_exile.database.data.spells.components.BaseFieldNeeder;
 import com.robertx22.age_of_exile.database.data.spells.components.MapHolder;
 import com.robertx22.age_of_exile.database.data.spells.map_fields.MapField;
+import com.robertx22.age_of_exile.database.data.spells.spell_classes.SpellCtx;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.util.math.BlockPos;
 
@@ -16,7 +17,7 @@ public abstract class BaseTargetSelector extends BaseFieldNeeder implements IGUI
         super(requiredPieces);
     }
 
-    public abstract List<LivingEntity> get(LivingEntity caster, LivingEntity target, BlockPos pos, MapHolder data);
+    public abstract List<LivingEntity> get(SpellCtx ctx, LivingEntity caster, LivingEntity target, BlockPos pos, MapHolder data);
 
     public static HashMap<String, BaseTargetSelector> MAP = new HashMap<>();