+++ src/main/java/com/robertx22/age_of_exile/database/data/spells/spell_classes/SpellModEnum.java
@@ -5,6 +5,7 @@ import com.robertx22.age_of_exile.uncommon.localization.Words;
 public enum SpellModEnum {
     CAST_SPEED(Words.CastSpeed, "cast_speed", -10),
     MANA_COST(Words.ManaCost, "mana_cost", -10),
+    AREA(Words.Normal_Gear, "area", -10),
     COOLDOWN(Words.Cooldown, "cooldown", -10),
     DAMAGE(Words.Damage, "damage", 5),
     HEALING(Words.Healing, "healing", 5),