+++ CHANGELOG.md
@@ -145,6 +145,8 @@ Or maybe 20 str + scaling over lvls slowly.
 * massive changes to internal stat calculation code, initially made to make the "Increased aura effect"  stat work. 
 Now I know the context of where each stat came from, so stats like "50% increased defenses from shields" can work!
 * added increased leech effect stat
+* new stat, spell area of effect damage
+* new stat, increased area of effect
 
 Spell adjustements:
 * whirlpool is now at caster, not a projectile